Fix incorrect reset() function during query plan step

+++ src/mongo/db/storage/key_string.h
@@ -398,7 +398,7 @@ public:
 
 private:
     void _appendRecordIdLong(int64_t val);
-    void _appendRecordIdStr(const char* str, int size);
+    void _appendRecordIdStr(const char* str, size_t size);
 
     void _appendAllElementsForIndexing(const BSONObj& obj,
                                        Ordering ord,
